Set up parcel and split out module
