+++ .gitignore
@@ -1,2 +1,6 @@
 node_modules
 .idea
+
+NOTES.md
+.parcel-cache
+dist
