+++ globals.d.ts
@@ -1,0 +1,7 @@
+import jquery from 'jquery';
+import highlightjs from 'highlightjs';
+
+declare global {
+    export const $: JQueryStatic;
+    export const hljs: typeof highlightjs;
+}
