+++ test/index.html
@@ -13,7 +13,7 @@
   <div id="qunit-fixture"></div>
   <script src="../node_modules/qunitjs/qunit/qunit.js"></script>
   <script src="../difflib.js"></script>
-  <script src="../codediff.js"></script>
+  <script src="../dist/codediff.js"></script>
   <script src="spanspreader_test.js"></script>
   <script src="htmlsubstr_test.js"></script>
   <script src="addcharacterdiffs_test.js"></script>
